fix: minor error using ray

+++ config.py
@@ -40,7 +40,7 @@ def args_for_tuning(parser):
     parser.add_argument('--n_trials', type=int, default=None, help='n_trials')
     parser.add_argument('--timeout', type=int, default=None, help='optuna training timeout(sec)')
     parser.add_argument('--n_job_parallel', type=int, default=1, help='n_job_parallel(optuna)')
-    parser.add_argument('--gpu_ratio', type=int, default=1, help='percentage of gpu resource for 1 trial(ray)')
+    parser.add_argument('--gpu_ratio', type=float, default=1, help='percentage of gpu resource for 1 trial(ray)')
 
 def args_for_config_file(parser):
     parser.add_argument('--config', default=None, type=str, help='read from config file')
